export makeProver and getQueryData
