+++ src/index.ts
@@ -2,6 +2,8 @@ export {
     DEFAULT_ALGORITHMS,
     DEFAULT_DIGESTS,
     DEFAULT_TRUST_ANCHORS,
+    getQueryData,
+    makeProver,
     dohQuery,
     DNSProver,
     NoValidDsError,
